Deleted old scraping system files, refactored code
