+++ src/main/java/sm/scraper/service/impl/ScraperServiceImpl.java
@@ -12,9 +12,7 @@ import sm.scraper.model.Price;
 import sm.scraper.repository.ItemRepository;
 import sm.scraper.repository.PriceRepository;
 import sm.scraper.service.ScraperService;
-import sm.scraper.util.NewScraper;
 import sm.scraper.util.Scraper;
-import sm.scraper.util.Scrapers;
 import sm.scraper.util.ScrapingConfig;
 
 import java.util.ArrayList;
@@ -38,13 +36,13 @@ public class ScraperServiceImpl implements ScraperService {
         String baseUrl = extractBaseUrl(itemUrl);
 
         // get scraper config from website_scraper_config.json (resources dir)
-        NewScraper newScraper = ScrapingConfig.getScraper(baseUrl);
-        if(newScraper == null) {
+        Scraper scraper = ScrapingConfig.getScraper(baseUrl);
+        if(scraper == null) {
             throw new EntityNotFoundException("Scraper for website " + baseUrl + " not implemented.");
         }
 
         // scrape item with full url provided
-        ItemDto itemDto = newScraper.scrape(itemUrl);
+        ItemDto itemDto = scraper.scrape(itemUrl);
 
         Item item = itemMapper.toEntity(itemDto);
         Price price = priceMapper.toEntity(itemDto.getCurrentPrice());
