--- src/main/java/sm/scraper/util/impl/AmazonScraper.java
@@ -1,12 +1,0 @@
-package sm.scraper.util.impl;
-
-import sm.scraper.dto.ItemDto;
-import sm.scraper.util.Scraper;
-
-public class AmazonScraper implements Scraper {
-    @Override
-    public ItemDto scrape(String url) {
-        System.out.println("scrapin AMAAAZON");
-        return null;
-    }
-}
